Revert "Moved add fighter, vehicle and gang additions to their own components"
